Removed AttachmentHub definitions, fixed folder name, refactored string extension method
