+++ src/dotnet/Common/Constants/Authorization/RoleDefinitionNames.cs
