+++ src/dotnet/Common/Constants/Authorization/RoleDefinitionNames.tt
